contextvault/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# HTTP Client
httpx==0.25.2
aiofiles==23.2.0

# CLI
click==8.1.7
rich==13.7.0

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# JSON & Data Processing
orjson==3.9.10

# Crypto & Security
cryptography==41.0.8
passlib[bcrypt]==1.7.4

# UUID
uuid==1.30

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Semantic Search & Embeddings
sentence-transformers==2.2.2
numpy==1.24.3
torch==2.1.0
transformers==4.35.0

# Dashboard & Web Interface
jinja2==3.1.2
python-multipart==0.0.6

# CLI Enhancements
click==8.1.7
rich==13.7.0
pyyaml==6.0.1

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
structlog==23.2.0
